Moved subroutine creation to import()
Subroutines created at compile time through the use of the implicit BEGIN
provided by 'use' would only exist in the first package that calls
Export::Lexical in a program.  The creation, and export, of the subroutines
has been moved to the import() subroutine to solve this.  While at it, the
Exporter module was removed, because it wasn't serving much of a purpose.