Removed superfluous non-numeric warning
The smart match operator was seeing the 0 and complaining when $_ was not
numeric, which it would be in the case of the "1,sub1,sub2" hint.  Changing
the "no" hint to an empty string cleared the warning and allowed the warn.t
test to pass (since it was confused by the warning).

Technically, removing the warnings pragma from lib/Export/Lexical.pm also
allowed warn.t to pass, but solving the problem correctly also works when -w
is given to perl itself.